requirements.txt
apache-airflow-providers-mysql
mysqlclient

